# packages/bytebot-agent/docs/OPENAPI_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: Browser-Use API
  description: |
    Enterprise-grade browser automation API with comprehensive session management, 
    task orchestration, and real-time monitoring capabilities.
    
    ## Features
    - **Local-Only Architecture**: 100% on-premises deployment with no cloud dependencies
    - **Enterprise Security**: JWT authentication with role-based access control  
    - **Real-time Monitoring**: Comprehensive health checks and performance metrics
    - **Browser Automation**: Screenshots, DOM interaction, form automation, data extraction
    - **Session Management**: Full lifecycle management of browser sessions
    - **Circuit Breaker Patterns**: Built-in reliability and resilience
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    Rate limits are enforced based on user roles:
    - ADMIN: 1000 requests/minute
    - OPERATOR: 500 requests/minute  
    - VIEWER: 300 requests/minute
    
  version: 2.0.0
  contact:
    name: Bytebot API Team
    url: https://docs.bytebot.ai/browser-use
    email: browser-use-support@bytebot.ai
  license:
    name: Enterprise License
    url: https://bytebot.ai/license
  termsOfService: https://bytebot.ai/terms
servers:
  - url: https://api.bytebot.ai/v1/browser-use
    description: Production Server
...
    description: Staging Server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /sessions:
    post:
      tags: [Sessions]
      summary: Create browser session
      description: |
        Create a new browser session with specified configuration and profile settings.
        Sessions provide isolated browser contexts for automation tasks.
      operationId: createBrowserSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBrowserSessionRequest'
            examples:
              basic_session:
                summary: Basic Session
                value:
                  name: "Data Extraction Session"
                  profile:
                    browserType: "chromium"
                    headless: true
              advanced_session:
                summary: Advanced Session with Custom Profile
                value:
                  name: "Advanced Automation Session"
                  description: "Session for complex automation tasks"
                  profile:
                    browserType: "chromium"
                    headless: false
                    windowWidth: 1920
                    windowHeight: 1080
                    userAgent: "Mozilla/5.0 (Custom Browser)"
                  timeoutSeconds: 900
                  enableScreenshots: true
                  enableVideoRecording: true
                  initialUrl: "https://example.com"
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Sessions]
      summary: List browser sessions
      description: Retrieve a list of browser sessions with their status and activity information
      operationId: listBrowserSessions
      security:
        - BearerAuth: []
      parameters:
        - name: active
          in: query
          description: Filter by active status
          required: false
          schema:
            type: boolean
        - name: status
          in: query
          description: Filter by session status
          required: false
          schema:
            type: string
            enum: [initializing, active, idle, closing, closed, error]
      responses:
        '200':
          description: Sessions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserSessionListResponse'

    delete:
      tags: [Sessions]
      summary: Close all browser sessions
      description: Close all active browser sessions and cleanup all resources
      operationId: closeAllBrowserSessions
      security:
        - BearerAuth: []
      responses:
        '204':
          description: All sessions closed successfully
  /sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session details
      description: Retrieve detailed information about a specific browser session
      operationId: getBrowserSession
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Unique identifier of the browser session
          schema:
            type: string
            pattern: '^session_[a-zA-Z0-9]{6,}$'
      responses:
        '200':
          description: Session details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserSessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Sessions]
      summary: Close browser session
      description: Close a specific browser session and cleanup all associated resources
      operationId: closeBrowserSession
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Unique identifier of the browser session
          schema:
            type: string
            pattern: '^session_[a-zA-Z0-9]{6,}$'
      responses:
        '204':
          description: Session closed successfully
        '404':
          $ref: '#/components/responses/NotFound'

  # Screenshot Endpoints
...
  /sessions/{sessionId}/type:
    post:
      tags: [DOM Operations]
      summary: Type text
      description: Type text into an input element identified by selector
      operationId: typeText
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Unique identifier of the browser session
          schema:
            type: string
            pattern: '^session_[a-zA-Z0-9]{6,}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeRequest'
      responses:
        '200':
          description: Text typed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserElementResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication for user sessions
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication for service-to-service calls
  schemas:
...
    CreateBrowserSessionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable name for the session
          example: "Data Extraction Session"
        description:
          type: string
          maxLength: 1000
          description: Description of the session purpose
        profile:
          $ref: '#/components/schemas/BrowserProfileConfig'
        timeoutSeconds:
          type: integer
          minimum: 60
          maximum: 3600
          default: 600
          description: Session timeout in seconds
        enableScreenshots:
          type: boolean
          default: true
          description: Enable screenshot capture capability
        enableVideoRecording:
          type: boolean
          default: false
          description: Enable video recording capability
        enableNetworkLogging:
          type: boolean
          default: true
          description: Enable request/response logging
        enableConsoleLogging:
          type: boolean
          default: true
          description: Enable console logging
        tags:
          type: array
          items:
            type: string
          description: Session-specific tags for organization
        initialUrl:
          type: string
          format: uri
          description: Initial URL to navigate to after session creation
        config:
          type: object
          additionalProperties: true
          description: Additional configuration parameters
...
    TypeRequest:
      type: object
      required:
        - selector
        - text
      properties:
        selector:
          type: string
          description: CSS selector for the input element
          example: "#search-input"
        text:
          type: string
          description: Text to type into the element
          example: "search query"
        clearFirst:
          type: boolean
          default: true
          description: Clear existing text before typing
        delay:
          type: integer
          minimum: 0
          maximum: 1000
          default: 0
          description: Delay between keystrokes in milliseconds
...
    BrowserSessionResponse:
      type: object
      required:
        - id
        - name
        - success
        - status
        - profile
        - createdAt
        - createdBy
      properties:
        id:
          type: string
          pattern: '^session_[a-zA-Z0-9]{6,}$'
          description: Unique session identifier
          example: "session_xyz789"
        name:
          type: string
          description: Session name
        description:
          type: string
          description: Session description
        success:
          type: boolean
          description: Whether session operation was successful
        status:
          type: string
          enum: [initializing, active, idle, closing, closed, error]
          description: Current session status
        profile:
          $ref: '#/components/schemas/BrowserProfileConfig'
        timeoutSeconds:
          type: integer
          description: Session timeout in seconds
        enableScreenshots:
          type: boolean
          description: Screenshot capture enabled
        enableVideoRecording:
          type: boolean
          description: Video recording enabled
        enableNetworkLogging:
          type: boolean
          description: Network logging enabled
        enableConsoleLogging:
          type: boolean
          description: Console logging enabled
        tags:
          type: array
          items:
            type: string
          description: Session tags
        createdAt:
          type: string
          format: date-time
          description: Session creation timestamp
        lastActivity:
          type: string
          format: date-time
          description: Last activity timestamp
        expiresAt:
          type: string
          format: date-time
          description: Session expiration timestamp
        createdBy:
          type: string
          description: User who created the session
        currentUrl:
          type: string
          format: uri
          description: Current URL in active tab
        currentTitle:
          type: string
          description: Current page title
        tabs:
          type: array
          items:
            $ref: '#/components/schemas/BrowserTabInfo'
          description: Open browser tabs
        activeTabsCount:
          type: integer
          description: Number of active tabs
        metrics:
          $ref: '#/components/schemas/BrowserSessionMetrics'
        processId:
          type: string
          description: Browser process ID
        websocketUrl:
          type: string
          format: uri
          description: WebSocket connection for real-time updates
        config:
          type: object
          additionalProperties: true
          description: Additional configuration
        error:
          type: object
          properties:
            message:
              type: string
              description: Error message
            code:
              type: string
              description: Error code
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
          description: Error information if session failed
    BrowserTabInfo:
      type: object
      required:
        - id
        - title
        - url
        - active
        - createdAt
        - lastActivity
        - loadingStatus
      properties:
        id:
          type: string
          description: Tab identifier
        title:
          type: string
          description: Tab title
        url:
          type: string
          format: uri
          description: Tab URL
        active:
          type: boolean
          description: Whether this tab is active
        createdAt:
          type: string
          format: date-time
          description: Tab creation timestamp
        lastActivity:
          type: string
          format: date-time
          description: Last activity timestamp
        loadingStatus:
          type: string
          enum: [loading, complete, error]
          description: Tab loading status
...
    BrowserSessionListResponse:
      type: object
      required:
        - sessions
        - total
        - activeCount
        - summary
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/BrowserSessionResponse'
          description: List of browser sessions
        total:
          type: integer
          description: Total number of sessions
        activeCount:
          type: integer
          description: Number of active sessions
        idleCount:
          type: integer
          description: Number of idle sessions
        errorCount:
          type: integer
          description: Number of sessions with errors
        totalMemoryUsageMB:
          type: number
          description: Total memory usage across all sessions in MB
        summary:
          type: object
          properties:
            totalSessions:
              type: integer
              description: Total sessions
            activeSessions:
              type: integer
              description: Active sessions
            idleSessions:
              type: integer
              description: Idle sessions
            closedSessions:
              type: integer
              description: Closed sessions
            averageSessionDuration:
              type: number
              description: Average session duration
            peakMemoryUsage:
              type: number
              description: Peak memory usage
          description: Summary statistics
...
    BrowserElementResponse:
      type: object
      required:
        - success
        - selector
        - action
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        selector:
          type: string
          description: CSS selector used for the operation
        action:
          type: string
          description: Action performed (click, type, etc.)
        result:
          type: string
          description: Result of the operation
        element:
          type: object
          properties:
            tagName:
              type: string
              description: HTML tag name
            text:
              type: string
              description: Element text content
            attributes:
              type: object
              additionalProperties:
                type: string
              description: Element attributes
            boundingBox:
              type: object
              properties:
                x:
                  type: number
                y:
                  type: number
                width:
                  type: number
                height:
                  type: number
              description: Element bounding box
          description: Element information
        timestamp:
          type: string
          format: date-time
          description: Operation timestamp
        executionTime:
          type: number
          description: Operation execution time in milliseconds
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
              timestamp: "2024-01-01T10:00:00Z"
              requestId: "req_abc123"
...
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "The requested resource was not found"
              timestamp: "2024-01-01T10:00:00Z"
...